Add missing bottom face for flowing liquids.
Draw the bottom face of a flowing liquid node if the node below is not
the same liquid, and not solid.

Change the texture to draw for top and bottom faces to the source
liquid texture when the liquid flows straight down, so the liquid
doesn't appear to flow in some random direction.

Include backface_culling flag in serialization format for TileDefs

This way flowing liquids actually show the backface when specified to
do so. Without this, TileDefs where by default initialized with
backface_culling = true and never set otherwise.

For backwards compatibility, an old client connected to a new server,
or a new client connected to an old server will behave like before
i.e., backface_culling is always true.

Draw face between liquid sources of different kinds

This is not the ideal way to do it. Ideally, we would take the
textures from both kinds of liquids, and show them on the opposite
sides, with backface_culling enabled. That would involve further
changes in the code, though.

With this patch, we take the texture from one liquid and show it on
both sides, i.e., with backface_culling disabled,